Add files via upload
These data were cleaned in the R Markdown file named, "TOC_2017incs_ForbesMoore.Rmd"; this file also includes some data visualization and exploration, and some learning of and exploring mathematical methods for analysis. The file named, "june2021_analyses.Rmd" includes more formal data exploration and analyses of the concentration data, as well as the carbon:nitrogen ratio for each feature.  The file named, "pool_calculations.Rmd" includes conversion of the concentration data to a) pool size per unit area, and also b) total pool size in each plot.  The visualizations in this script are the final ones, and the models are all the final models (explored in "june2021_analyses").